Agentic_AI_with_memory/requirements.txt
# Core dependencies
langchain>=0.1.0
langchain-openai>=0.1.0
langchain-core>=0.2.40
langgraph>=0.1.0
pydantic>=2.0.0
python-dotenv>=1.0.0

# Knowledge Graph
stardog

# Data processing
tabulate

# Utilities
typing-extensions